Move agent pre- and post-callbacks to separate module
